feat(user_books): add user_books model
This allows for a many-to-many relationship between books and users,
also storing information about the user's relation to the book, such as
rating, when they began reading etc. However, we don't store information
such as "read page x on date y", as a design choice to keep the
application simple.